using session.isValid  to check for a valid session before sending to not_authorized page, also no  longer using  window.location in favor of history.pushState

+++ pic-sure-hpds-ui/src/main/webapp/picsureui/common/transportErrors.js
@@ -1,14 +1,15 @@
-define(["jquery"],
-    function($) {
-        var transportErrorHandlers = {};
-        transportErrorHandlers.redirectionUrl = "/psamaui/login?redirection_url=" + "/picsureui/";
+define(["jquery", "studyAccess/studyAccess", "common/session"],
+    function($, studyAccess, session) {
+        var transportErrorHandlers = {
+            session: session
+        };
 
-        transportErrorHandlers.handleAll = function (response, logMsg) {
-            console.debug(logMsg);
-            console.dir(response);
+        transportErrorHandlers.redirectionUrl = "/psamaui/login?redirection_url=/picsureui/";
+
+        transportErrorHandlers.handleAll = function (response) {
             var hasError = false;
-            // list of all error handlers here
             if (this.handle401(response)) {
+                console.debug("Captured HTTP 401 response");
                 hasError = true;
             }
             return hasError;
@@ -18,9 +19,12 @@ define(["jquery"],
         transportErrorHandlers.handle401 = function (response, redirectionUrl = false) {
             if (redirectionUrl === false) redirectionUrl = this.redirectionUrl;
             if (response.status === 401) {
-                sessionStorage.clear();
-                window.location = redirectionUrl;
-                return true;
+                if (this.session.isValid()) {
+                    sessionStorage.clear();
+                    history.pushState({}, "Login", redirectionUrl);
+                } else {
+                    history.pushState({}, "Not Authorized", "/psamaui/not_authorized");
+                }
             }
             return false;
         }.bind(transportErrorHandlers);
